Move get_free_port and debug functions into utils.py
